+++ src/BT/ActionNode.hpp
@@ -142,6 +142,8 @@ protected:
     virtual State::StateType performAction() override;
     virtual State::StateType continueAction() override;
 
+    virtual void getLuaStatesHelper(std::vector<lua_State*>& v) override;
+
 private:
     State::StateType performLuaScript(bool isContinuing);