+++ src/BT/BehaviorNode.cpp
@@ -112,3 +112,20 @@ BT::BehaviorNode* BT::BehaviorNode::findByID(std::string id)
     }
 }
 
+std::vector<lua_State*> BT::BehaviorNode::getLuaStatesInTree()
+{
+    std::vector<lua_State*> states;
+
+    getLuaStatesHelper(states);
+
+    return states;
+}
+
+void BT::BehaviorNode::getLuaStatesHelper(std::vector<lua_State*>& v)
+{
+    for(auto& child : children)
+    {
+        child->getLuaStatesHelper(v);
+    }
+}
+